Improve handling of IP address config
When detecting the IP address of the control planes check
if disable_ipv4 / disable_ipv6 variables were set.

This is needed in an update scenario when IPv4 / IPv6 is
disabled for existing control planes. The Terraform state
still contains the IP address but the connection is no longer
possible.

Also wrap the IPv6 addresses in square brackets to prevent
"too many colons in address" errors in SSH commands.